Crate nuthru

The crate nathru (NUmber THeory in RUst) implements a few simple number theory functions.

Functions

cornacchia
Finds integer solution to x2 + y2 = p. See https://en.wikipedia.org/wiki/Cornacchia's_algorithm.

gcd
The greatest common divisor of a and b

int_sqrt
The integer square root

legendre_symbol
Legendre symbol, a multiplicative function with 1, 0, or -1 mod p. Spcifically, its value on a (nonzero) quadratic residue mod p is 1 and on a non-quadratic residue (non-residue) is −1. Its value on zero is 0. See https://en.wikipedia.org/wiki/Legendre_symbol.

modulo
This function complements the % operator, as it behaves like the normal mathematical mod operator even for negative inputs.

power_mod
Computes powers modulo m

tonelli_shanks
The Tonelli–Shanks algorithm finds solutions to x2 = n mod p, where p is an odd prime.